backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.0.3
redis==5.0.1
celery==5.3.4
requests==2.31.0
python-telegram-bot==20.7
asyncpg==0.29.0
pytest==7.4.3
pytest-asyncio==0.21.1
aiohttp==3.9.1
structlog==23.2.0 